BigDataAplicado/ConsumoDeDatos/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dateutil==2.8.2
streamlit==1.29.0
plotly==5.18.0
streamlit-autorefresh==0.0.1
